+++ src/main/java/controller/springcontroller/DirectMessage.java
@@ -2,8 +2,11 @@ package controller.springcontroller;
 
 import java.io.IOException;
 import java.io.PrintWriter;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.TreeMap;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -13,17 +16,22 @@ import javax.sound.midi.Soundbank;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.google.gson.Gson;
 
 import model.UserModel;
 import model.springmodel.Message;
+import model.springmodel.Question;
+import notifier.ServerEndPoint;
+import notifier.SessionId;
 import service.springservice.DirectMessageService;
 
 @Controller
@@ -36,7 +44,7 @@ public class DirectMessage
 		
 	
 	@PostMapping("/sendDM")
-	public String sendDM(@ModelAttribute ("message") Message themessage, HttpServletRequest request,Model theModel,@RequestParam (name="id",required=false) String receiverid,@RequestParam (name="name",required=false) String receivername)
+	public String sendDM(@ModelAttribute ("message") Message themessage, HttpServletRequest request,Model theModel,@RequestParam (name="id",required=false) String receiverid,@RequestParam (name="name",required=false) String receivername,RedirectAttributes redirectAttributes)
 	{	
 		HttpSession session=request.getSession();
 
@@ -58,7 +66,7 @@ public class DirectMessage
 	
 	
 	@GetMapping("/inbox")
-	public String myMessages(HttpServletRequest request, Model theModel, @RequestParam (name="id",required=false) String receiverid,@RequestParam (name="name",required=false) String threadname  )
+	public String myMessages(HttpServletRequest request, Model theModel, @RequestParam (name="id",required=false) String receiverid,@RequestParam (name="name",required=false) String threadname)
 	{
 		HttpSession session=request.getSession();
 
@@ -67,11 +75,28 @@ public class DirectMessage
 		Object object=session.getAttribute("userModel");
 		String senderid=um.getUserId(object);
 		
-		HashMap<String, String> theThreads= dmservice.getMessageThreads(senderid);
+		TreeMap<String, String> theThreads= dmservice.getMessageThreads(senderid);
+		List<SessionId> allUsers=ServerEndPoint.users;
+		SessionId isUser=null;
+		ArrayList<Boolean> onlineUsers=new ArrayList<>();
 		
+		for(Map.Entry<String, String> entries : theThreads.entrySet())
+		{
+			isUser=new SessionId(null,entries.getKey());
+			
+			if(allUsers.contains(isUser))
+			{
+				onlineUsers.add(true);
+				
+			}
+			
+			else
+			{
+				onlineUsers.add(false);
+			}
+		}
 		theModel.addAttribute("threads", theThreads);
-		
-		
+		theModel.addAttribute("onlineUsers",onlineUsers);	
 		//conversation
 				
 		if(receiverid!=null)
