+++ src/main/java/dao/springdao/DirectMessageDAO.java
@@ -2,6 +2,7 @@ package dao.springdao;
 
 import java.util.HashMap;
 import java.util.List;
+import java.util.TreeMap;
 
 import model.UserModel;
 import model.springmodel.Message;
@@ -14,7 +15,7 @@ public interface DirectMessageDAO {
 	List<Message> showConversation(String receiverid, String senderid);
 
 
-	HashMap<String, String> getMessageThreads(String id);
+	TreeMap<String, String> getMessageThreads(String id);
 
 
 	List<UserModel> searchThreadName(String name);
